refactor: switch to namespace
