+++ .eslintrc.json
@@ -34,6 +34,7 @@
     "import/prefer-default-export": "off",
     "react/prop-types": "off",
     "@typescript-eslint/explicit-function-return-type": "off",
+    "@typescript-eslint/triple-slash-reference": "off",
     "import/extensions": [
       "error",
       "ignorePackages",
